requirements.txt
pytest==3.4.0
GitPython==2.1.8
PyGithub==1.36
